Christopher Graf - 05/11/21
I had a very strange bug that was doubling files in the Project window and causing lots of harm. I had to delete and undo many things. I finally got it back to normal, but lost a lot of framework. Making this commit to know it is working again and make new commits for the submission process