Christopher Graf - 05/11/21 - Secondary Fire Powerup
Created a new powerup that has a 1 in 10 chance of spawning each time any powerup is spawning. The powerup creates a new fire system where lasers shoot from the player in 4 directions. Straight up, down, left and right. Changed the Laser script so that the parent holding multiple lasers will only be destroyed if they are the last child. Also have them destroy themselves if they are out of bounds in any direction.